moving pairwise tests to own file, adding xml read also
