+++ integration_tests/test_rough_align.py
@@ -25,10 +25,6 @@ from rendermodules.module.render_module import RenderModuleException
 from rendermodules.materialize.render_downsample_sections import RenderSectionAtScale, create_tilespecs_without_mipmaps
 from rendermodules.dataimport.make_montage_scapes_stack import MakeMontageScapeSectionStack, create_montage_scape_tile_specs
 from rendermodules.rough_align.do_rough_alignment import SolveRoughAlignmentModule
-from rendermodules.rough_align.pairwise_rigid_rough import \
-        PairwiseRigidRoughAlignment
-from rendermodules.rough_align.make_anchor_stack import \
-        MakeAnchorStack
 from rendermodules.solver.solve import Solve_stack
 from rendermodules.rough_align.apply_rough_alignment_to_montages import (ApplyRoughAlignmentTransform,
                                                                          #example as ex1,
@@ -1283,298 +1279,3 @@ def test_filterNameList_warning_rendersection(tmpdir):
         "filterListName": "notafilter"}
     with pytest.warns(UserWarning):
         mod = RenderSectionAtScale(input_data=warn_input, args=[])
-
-
-def test_make_anchor_stack(
-        render,
-        montage_scape_stack,
-        tmpdir_factory):
-
-    zs = renderapi.stack.get_z_values_for_stack(
-            montage_scape_stack,
-            render=render)
-
-    # make some transforms that would come from a manual trakem alignment
-    transforms = {}
-    for z in zs[:-1]:
-        tf = renderapi.transform.AffineModel(
-                M00=1.0 + np.random.randn(),
-                M11=1.0 + np.random.randn())
-        transforms["%d_blah.png" % z] = tf.to_dict()
-    output_directory = str(tmpdir_factory.mktemp('anchor_stack_input'))
-    tjson_path = os.path.join(output_directory, 'human_made_transforms.json')
-    with open(tjson_path, 'w') as f:
-        json.dump(transforms, f, indent=2)
-
-    anchor_stack = "my_anchor_stack"
-    ex = {
-            "render": dict(solver_example['render']),
-            "input_stack": montage_scape_stack,
-            "output_stack": anchor_stack,
-            "transform_json": tjson_path,
-            "close_stack": True
-            }
-    mmod = MakeAnchorStack(input_data=ex, args=[])
-    mmod.run()
-
-    tspecs = renderapi.tilespec.get_tile_specs_from_stack(
-            ex['output_stack'], render=render)
-    for t in tspecs:
-        assert len(t.tforms) == 1
-        for tf in transforms:
-            if tf.startswith('%d_' % t.z):
-                assert t.tforms[0].dataString == transforms[tf]['dataString']
-
-    # pedantic, what if a downsample stack has multiple tiles for one z
-    other_stack = "other_stack"
-    renderapi.stack.clone_stack(montage_scape_stack, other_stack, render=render)
-    tspecs = renderapi.tilespec.get_tile_specs_from_stack(
-            other_stack, render=render)
-    t = tspecs[0]
-    t.tileId = 'something_unique'
-    renderapi.client.import_tilespecs(other_stack, [t], render=render)
-    ex['input_stack'] = other_stack
-    mmod = MakeAnchorStack(input_data=ex, args=[])
-    with pytest.raises(RenderModuleException):
-        mmod.run()
-    renderapi.stack.delete_stack(other_stack, render=render)
-
-    # add a z to the transforms that isn't in the stack
-    ztry = zs[-1] + 1
-    while ztry in zs:
-        ztry += 1
-    tf = renderapi.transform.AffineModel(
-                M00=1.0 + np.random.randn(),
-                M11=1.0 + np.random.randn())
-    transforms["%d_blah.png" % ztry] = tf.to_dict()
-    with open(tjson_path, 'w') as f:
-        json.dump(transforms, f, indent=2)
-    ex['input_stack'] = montage_scape_stack
-    mmod = MakeAnchorStack(input_data=ex, args=[])
-    with pytest.raises(renderapi.errors.RenderError):
-        mmod.run()
-    renderapi.stack.delete_stack(ex['output_stack'], render=render)
-
-
-def calc_residuals(tspecs, matches):
-    tids = np.array([t.tileId for t in tspecs])
-    res = []
-    for m in matches:
-        pind = np.argwhere(tids == m['pId']).flatten()[0]
-        qind = np.argwhere(tids == m['qId']).flatten()[0]
-        p = np.array(m['matches']['p']).transpose()
-        q = np.array(m['matches']['q']).transpose()
-        for tf in tspecs[pind].tforms:
-                p = tf.tform(p)
-        for tf in tspecs[qind].tforms:
-                q = tf.tform(q)
-        res += np.linalg.norm(p - q, axis=1).tolist()
-    return res
-
-
-def test_pairwise_results(render, tmpdir_factory):
-    render_params['project'] = 'rough_align_test'
-    # ground truth tilespecs, some gradual rotation and translation
-    nz = 10
-    z = np.arange(nz)
-    theta = np.arange(nz) * 0.002
-    dx = np.arange(nz) * 100 + 245
-    dy = np.arange(nz) * -200 - 356
-    ground_truth = [renderapi.tilespec.TileSpec() for i in np.arange(nz)]
-    downsample = [renderapi.tilespec.TileSpec() for i in np.arange(nz)]
-    # downsample stack will have default affine
-    for i in range(nz):
-        c, s = np.cos(theta[i]), np.sin(theta[i])
-        tf = renderapi.transform.AffineModel(
-                M00=c, M01=-s, M10=s, M11=c, B0=dx[i], B1=dy[i])
-        ground_truth[i].tforms = [
-                renderapi.transform.AffineModel(json=tf.to_dict())]
-        downsample[i].tforms = [renderapi.transform.AffineModel()]
-        for tspec in [ground_truth[i], downsample[i]]:
-            tspec.ip['0'] = renderapi.image_pyramid.MipMap()
-            tspec.z = z[i]
-            tspec.layout.sectionId = str(float(z[i]))
-            tspec.tileId = '%d' % z[i]
-            tspec.width = 1500
-            tspec.height = 1500
-
-    # point matches are image-based, i.e. ground truth
-    # for pairwise rigid, we need them between the ground truth pairs
-    matches = []
-    for i in range(nz - 1):
-        src = np.random.rand(10, 2) * 1500
-        p = ground_truth[i].tforms[0].inverse_tform(src)
-        q = ground_truth[i + 1].tforms[0].inverse_tform(src)
-        w = [1.0] * 10
-        matches.append({
-            "pGroupId": ground_truth[i].layout.sectionId,
-            "qGroupId": ground_truth[i + 1].layout.sectionId,
-            "pId": ground_truth[i].tileId,
-            "qId": ground_truth[i + 1].tileId,
-            "matches": {
-                "w": w,
-                "p": p.transpose().tolist(),
-                "q": q.transpose().tolist()}})
-
-    # make the downsample stack and collection
-    ds_stack = "pw_downsample_stack"
-    ds_coll = "pw_downsample_collection"
-    renderapi.stack.create_stack(ds_stack, render=render)
-    renderapi.client.import_tilespecs(ds_stack, downsample, render=render)
-    renderapi.stack.set_stack_state(ds_stack, state='COMPLETE', render=render)
-    renderapi.pointmatch.import_matches(ds_coll, matches, render=render)
-
-    # a human determines the true relative transforms for 2 sections
-    transforms = {}
-    transforms['%d_blah' % z[2]] = ground_truth[2].tforms[0].to_dict()
-    transforms['%d_blah' % z[-1]] = ground_truth[-1].tforms[0].to_dict()
-    output_directory = str(tmpdir_factory.mktemp('pairwise'))
-    tjson_path = os.path.join(output_directory, 'human_made_transforms.json')
-    with open(tjson_path, 'w') as f:
-        json.dump(transforms, f, indent=2)
-
-    # make an anchor stack from that
-    anchor_stack = "an_anchor_stack"
-    ex = {
-            "render": dict(render_params),
-            "input_stack": ds_stack,
-            "output_stack": anchor_stack,
-            "transform_json": tjson_path,
-            "close_stack": True
-            }
-    mmod = MakeAnchorStack(input_data=ex, args=[])
-    mmod.run()
-
-    aligned_stack = "aligned_stack"
-    pw_example = {
-            "render": dict(render_params),
-            "input_stack": ds_stack,
-            "output_stack": aligned_stack,
-            "anchor_stack": anchor_stack,
-            "match_collection": ds_coll,
-            "close_stack": True,
-            "minZ": z[0],
-            "maxZ": z[-1],
-            "gap_file": None,
-            "pool_size": 2,
-            "output_json": os.path.join(output_directory, 'output.json'),
-            "overwrite_zlayer": True,
-            "translate_to_positive": True,
-            "translation_buffer": [50, 50]
-            }
-    pwmod = PairwiseRigidRoughAlignment(input_data=dict(pw_example), args=[])
-    pwmod.run()
-
-    with open(pwmod.args['output_json'], 'r') as f:
-        pwout = json.load(f)
-
-    # check the residuals
-    new_specs = renderapi.tilespec.get_tile_specs_from_stack(
-            pwout['output_stack'], render=render)
-    res_new = calc_residuals(new_specs, matches)
-    assert np.all(np.array(res_new) < 1.0)
-
-    # everything rotated correctly?
-    assert np.all(np.isclose(
-        theta,
-        np.array([t.tforms[0].rotation for t in new_specs])))
-
-    # in this artificial example, the ground truth is changing
-    # in a smooth linear way, we can check the results are 
-    # weighted linear averages of the anchors
-    for i in range(2, 9):
-        assert np.all(
-                np.isclose(
-                    new_specs[i].tforms[0].M,
-                    (new_specs[2].tforms[0].M * (7 - np.abs(2 - i)) +
-                        new_specs[9].tforms[0].M * (7 - np.abs(9 - i)))/ 7,
-                    1e-3))
-
-    # finally, the module gives good residuals without an anchor stack
-    # one should just be careful doing this over long stretches
-    # as it may lead to some undesired, unphysical long-range drift
-    # in the solution
-    pw_example['anchor_stack'] = None
-    pwmod = PairwiseRigidRoughAlignment(input_data=dict(pw_example), args=[])
-    pwmod.run()
-    with open(pwmod.args['output_json'], 'r') as f:
-        pwout = json.load(f)
-    # check the residuals
-    new_specs = renderapi.tilespec.get_tile_specs_from_stack(
-            pwout['output_stack'], render=render)
-    res_new = calc_residuals(new_specs, matches)
-    assert np.all(np.array(res_new) < 1.0)
-
-    # delete the render things
-    renderapi.stack.delete_stack(ds_stack, render=render)
-    renderapi.stack.delete_stack(anchor_stack, render=render)
-    renderapi.stack.delete_stack(pwout['output_stack'], render=render)
-    renderapi.pointmatch.delete_collection(ds_coll, render=render)
-
-
-def test_pairwise_rigid_alignment_coverage(
-        render,
-        montage_scape_stack,
-        rough_point_match_collection,
-        tmpdir_factory):
-    output_lowres_stack = '{}_DS_Pairwise_rigid'.format(montage_scape_stack)
-
-    output_directory = str(tmpdir_factory.mktemp('pwr_output_json'))
-
-    example = {
-            "render": dict(solver_example['render']),
-            "input_stack": montage_scape_stack,
-            "output_stack": output_lowres_stack,
-            "match_collection": rough_point_match_collection,
-            "close_stack": True,
-            "minZ": 1020,
-            "maxZ": 1022,
-            "gap_file": None,
-            "pool_size": 2,
-            "output_json": os.path.join(output_directory, 'output.json'),
-            "overwrite_zlayer": True,
-            "translate_to_positive": True,
-            "translation_buffer": [50, 50]
-            }
-
-    # normal
-    ex = dict(example)
-    pwmod = PairwiseRigidRoughAlignment(input_data=ex, args=[])
-    pwmod.run()
-    with open(ex['output_json'], 'r') as f:
-        outj = json.load(f)
-    assert(outj['masked'] == [])
-    assert(outj['missing'] == [])
-    assert(len(outj['residuals']) == 2)
-
-    # gap file
-    ex = dict(example)
-    gapfile = os.path.join(output_directory, "gapfile.json")
-    with open(gapfile, 'w') as f:
-        json.dump({"1020": "gap for some reason"}, f)
-    ex['gap_file'] = gapfile
-    pwmod = PairwiseRigidRoughAlignment(input_data=ex, args=[])
-    pwmod.run()
-    with open(ex['output_json'], 'r') as f:
-        outj = json.load(f)
-    assert(outj['masked'] == [])
-    assert(outj['missing'] == [1020])
-    assert(len(outj['residuals']) == 1)
-
-    # make one match missing
-    ex = dict(example)
-    render = renderapi.connect(**ex['render'])
-    matches = renderapi.pointmatch.get_matches_outside_group(
-            rough_point_match_collection,
-            "1020.0",
-            render=render)
-    new_collection = 'with_missing'
-    renderapi.pointmatch.import_matches(
-            new_collection,
-            matches,
-            render=render)
-    ex['match_collection'] = new_collection
-
-    with pytest.raises(RenderModuleException):
-        pwmod = PairwiseRigidRoughAlignment(input_data=ex, args=[])
-        pwmod.run()
